+++ libnw/Cargo.toml
@@ -4,3 +4,4 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+micromath = { version = "2.1.0", features = ["vector"] }
